fix compiler warning

+++ symmetric-client-clib/src/util/Base64.c
@@ -89,10 +89,10 @@ unsigned char * SymBase64_decode(const char *data, int *outSize) {
 
     int i, j;
     for (i = 0, j = 0; i < inSize;) {
-        uint32_t sextet_a = data[i] == '=' ? 0 & i++ : decodingTable[data[i++]];
-        uint32_t sextet_b = data[i] == '=' ? 0 & i++ : decodingTable[data[i++]];
-        uint32_t sextet_c = data[i] == '=' ? 0 & i++ : decodingTable[data[i++]];
-        uint32_t sextet_d = data[i] == '=' ? 0 & i++ : decodingTable[data[i++]];
+        uint32_t sextet_a = data[i] == '=' ? 0 & i++ : decodingTable[(int) data[i++]];
+        uint32_t sextet_b = data[i] == '=' ? 0 & i++ : decodingTable[(int) data[i++]];
+        uint32_t sextet_c = data[i] == '=' ? 0 & i++ : decodingTable[(int) data[i++]];
+        uint32_t sextet_d = data[i] == '=' ? 0 & i++ : decodingTable[(int) data[i++]];
 
         uint32_t triple = (sextet_a << 3 * 6) + (sextet_b << 2 * 6) + (sextet_c << 1 * 6) + (sextet_d << 0 * 6);
